update
aaa

+++ DATA/Models/CentrodeCosto.cs
@@ -8,7 +8,7 @@ namespace DATA.Models
     {
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public int idCentrodeCosto { get; set; }
+        public long idCentrodeCosto { get; set; }
         [MaxLength(50)]
         public string CentroDeCosto { get; set; }
         public string Obs { get; set; }
